Enrich reserr monad with a list of non fatal warnings
The monad is now the pair of what it was and a list of warnings.
We also add a promote and a map functions to collect information from a
list of data from which we can build a reserr. The level function is
used in the promote process in order to decide whether the warning is
fatal or not.

Co-authored-by: Samuel Hym <samuel.hym@rustyne.lautre.net>