Make postcond return the reason of failure rather than a bool
The Spec module contains a dummy postcond function, always returning
true. The new postcond function return an `(string * [(string *
location)]) option` where the first string is the command name and the
second one the term being violated. Checking one term return a value of
such a type, with a singleton for the list. Results of checking several
terms will be combine using an append function defined in the runtime.